Add last run time route to multiple_ukrdcids
