+++ ukrdc_fastapi/routers/api/v1/admin/__init__.py
@@ -1,4 +1,5 @@
 import datetime
+from sys import prefix
 from typing import Optional
 
 from fastapi import APIRouter, Depends, Security
@@ -8,13 +9,15 @@ from ukrdc_sqla.stats import PatientsLatestErrors
 
 from ukrdc_fastapi.dependencies import get_jtrace, get_statsdb
 from ukrdc_fastapi.dependencies.auth import Permissions, UKRDCUser, auth
-from ukrdc_fastapi.query.stats import get_full_errors_history, get_multiple_ukrdcids
+from ukrdc_fastapi.query.stats import get_full_errors_history
 from ukrdc_fastapi.query.workitems import get_full_workitem_history, get_workitems
-from ukrdc_fastapi.schemas.admin import AdminCountsSchema, MultipleUKRDCIDGroup
+from ukrdc_fastapi.schemas.admin import AdminCountsSchema
 from ukrdc_fastapi.schemas.common import HistoryPoint
-from ukrdc_fastapi.utils.paginate import Page, paginate_sequence
+
+from . import datahealth
 
 router = APIRouter(tags=["Admin"])
+router.include_router(datahealth.router, prefix="/datahealth")
 
 
 @router.get(
@@ -100,28 +103,3 @@ def admin_counts(
         UKRDC_records=ukrdc_records_count,
         patients_receiving_errors=patients_receiving_errors_count,
     )
-
-
-# Data health reports
-
-
-@router.get(
-    "/datahealth/multiple_ukrdcids",
-    response_model=Page[MultipleUKRDCIDGroup],
-    dependencies=[
-        Security(
-            auth.permission(
-                [
-                    Permissions.READ_RECORDS,
-                    Permissions.UNIT_PREFIX + Permissions.UNIT_WILDCARD,
-                ]
-            )
-        )
-    ],
-)
-def datahealth_multiple_ukrdcids(
-    jtrace: Session = Depends(get_jtrace),
-    statsdb: Session = Depends(get_statsdb),
-):
-    """Retreive list of patients with multiple UKRDC IDs"""
-    return paginate_sequence(get_multiple_ukrdcids(statsdb, jtrace))
